Refactor code and update UI in Router.jsx, HomeScreen.jsx, LoginPage.jsx, RegisterPage.jsx, SearchPage.jsx
