--- front/src/components/ButtonComponent.jsx
@@ -1,5 +1,0 @@
-const ButtonComponent = ({text}) => (
-    <button>{text}</button>
-    );
-
-export default ButtonComponent;
